requirements.txt
PyPDF2==3.0.1
pandas==2.1.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
tqdm==4.66.1
Flask==3.0.2
Werkzeug==3.0.1
gunicorn==21.2.0 